requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.20
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
importlib-metadata==6.7.0
Markdown==3.4.4
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
zipp==3.15.0
